// src/v2/front/InSeconds.Client/src/app/api/catalogue/api.generated.ts
export interface DeezerTrackResult {
    deezerTrackId: number;
    artist: string;
    title: string;
    previewUrl: string | undefined;
    coverUrl: string | undefined;
    releaseYear: number | undefined;
    rank: number | undefined;

    [key: string]: any;
}
// ...
export interface TrackListItem {
    id: number;
    deezerTrackId: number;
    artist: string;
    title: string;
    coverUrl: string | undefined;
    releaseYear: number | undefined;
    rank: number | undefined;
    previewStatus: string;
    previewCheckedAt: Date | undefined;
    isDisabled: boolean;
    lastUsedDate: Date | undefined;
    usageCount: number;
    unlockDate: Date | undefined;
    inTodayChallenge: boolean;

    [key: string]: any;
}